+++ api/bootstrap/load_modules.R
@@ -159,6 +159,7 @@ bootstrap_load_modules <- function() {
     "functions/response-fields-helpers.R",
     "functions/email-templates.R",
     "functions/pagination-helpers.R",
+    "functions/external-proxy-request-state.R",
     "functions/external-proxy-functions.R",
     "functions/external-proxy-gnomad.R",
     "functions/external-proxy-gnomad-batch.R",